refactor: dto 이름 변경
